dbx_workflow/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
# "raw-ingest" pour matcher le job configuration
name = "raw-ingest"
version = "0.0.1"
description = "Ticker raw data ingestion and ETL pipeline"
readme = "README.md"
requires-python = ">=3.10,<3.13"
authors = [{name = "Mickael", email = "mickael.rakotoa@gmail.com"}]

dependencies = [
    "databricks>=0.2",
    "databricks-connect>=15.4.17",
    "python-dotenv",
    "pytest>=9.0.2",
    "pydantic>=2.0.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "databricks-dlt",
    "databricks-connect>=15.4,<15.5",
    "datamodel-code-generator[http]>=0.25.0",
    "genson",
]

# Le package est dans src/, pas à la racine
[tool.hatch.build.targets.wheel]
packages = ["src/raw_ingest"]

# Pas de "src." au début du path
[project.scripts]
ingest_market_price_data = "raw_ingest.ingest_market_price_data:main"
ingest_technical_indicators = "raw_ingest.ingest_technical_indicators:main"
ingest_swapi_characters = "raw_ingest.ingest_swapi_characters:main"

[tool.black]
line-length = 125

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]
